Resolve endianess issues

+++ hqc/patches/Reference_Implementation_hqc-192_src_repetition.c
@@ -27,7 +27,7 @@
 +	uint32_t pos_r;
 +	uint64_t *p64 = em;
 +	const uint64_t mask[2][2] = {{0x0UL,0x0UL}, {0x7FFFFFFFFFFFFFFUL,0x3FFFFFFFFFFFFFFUL}};
-+	for (i = 0 ; i < VEC_N1_SIZE_64 - 1 ; i++) {
++	for (i = 0 ; i < (uint16_t) (VEC_N1_SIZE_64 - 1) ; i++) {
 +		for (j = 0 ; j < 64 ; j++) {
 +			bit = (m[i] >> j) & 0x1;
 +			pos_r = PARAM_N2 * ((i << 6) + j);
