+++ hqc/patches/Reference_Implementation_hqc-rmrs-128_src_code.h
@@ -6,8 +6,8 @@
  
 -void code_encode(uint64_t *codeword, const uint64_t *message);
 -void code_decode(uint64_t *message, const uint64_t *vector);
-+void code_encode(uint64_t *em, const uint64_t *message);
-+void code_decode(uint64_t *m, const uint64_t *em);
++void code_encode(uint8_t *em, const uint8_t *message);
++void code_decode(uint8_t *m, const uint8_t *em);
  
  #endif
